restructure xml extension
