+++ src/input/SConscript
@@ -14,6 +14,9 @@
 Import('env')
 
 env.sourceFiles.append(env.Object('ParallelVertexFilter.cpp'))
+env.sourceFiles.append(env.Object('AnalysisAttributes.cpp'))
+env.sourceFiles.append(env.Object('MeshAttributes.cpp'))
+env.sourceFiles.append(env.Object('split.cpp'))
 env.sourceFiles.append(env.Object('../../../submodules/tinyxml2/tinyxml2.cpp'))
 
 if env['use_netcdf']:
