+++ src/input/split.h
@@ -1,0 +1,18 @@
+#ifndef MYSPLIT_H
+#define MYSPLIT_H
+
+#include <algorithm>
+#include <cstring>
+#include <string>
+
+#include <list>
+#include <iostream>
+#include <fstream>
+#include <vector>
+
+
+std::vector<std::string> &split(std::vector<std::string> &elems, const std::string &s, char delim);
+
+
+
+#endif
